Module background_hang_monitor::background_hang_monitor

Structs

BackgroundHangMonitorChan 🔒

A wrapper around a sender to the monitor, which will send the Id of the monitored component along with each message, and keep track of whether the monitor is still listening on the other end.

BackgroundHangMonitorWorker 🔒

HangMonitorRegister

MonitoredComponent 🔒

Sample 🔒

SignalToExitOnDrop 🔒

Wraps BackgroundHangMonitorExitSignal and calls signal_to_exit when dropped.

Enums

MonitoredComponentMsg 🔒

Messages sent from monitored components to the monitor.

Never 🔒

Stable equivalent to the ! type

Type Aliases

MonitoredComponentReceiver 🔒

MonitoredComponentSender 🔒